requirements.txt
discord.py[voice]==2.1.0
python-dotenv==0.21.0
tinytag==1.8.1
requests==2.28.1
Pillow==9.4.0
